POS_SYSTEM/pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "FUT 2024 playground."
authors = ["FUT Dev <dev@freeuni.edu.ge>"]
readme = "final.md"

[tool.poetry.dependencies]
python = "^3.11"
coverage = "^7.6.1"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
pytest-vcr = "^1.0.2"
apexdevkit = "^1.16.3"
faker = "^33.1.0"
setuptools = "~=69.0.3"
requests = "~=2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "*"
hypothesis = "^6.122.4"
faker = "^33.1.0"
behave = "^1.2.6"
types-requests = "^2.32.0.20241016"

[tool.poetry.group.lint.dependencies]
mypy = "*"
ruff = "*"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true

[tool.ruff]
target-version = "py311"

lint.select = ["E", "F", "I"]
lint.ignore = []
lint.fixable = ["A", "B", "C", "D", "E", "F", "I"]
lint.unfixable = []

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.coverage.run]
branch = true
source = []

[tool.coverage.report]
skip_empty = true
skip_covered = true
show_missing = true
